+++ src/renderer/src/constants.ts
@@ -1,6 +1,6 @@
 import { Downloader } from "@shared";
 
-export const VERSION_CODENAME = "Aqua";
+export const VERSION_CODENAME = "Breakwater";
 
 export const DOWNLOADER_NAME = {
   [Downloader.RealDebrid]: "Real-Debrid",